requirements.txt
fastapi
uvicorn
requests
paho-mqtt